+++ bot-tr/Program.cs
@@ -24,7 +24,7 @@ class Program
 
         using CancellationTokenSource cts = new();
         var botToken = "6872155143:AAEVSYg6avTl15uqB650WKM6-u30XqOuR5s";
-        var botHandler = new BotHandler(botToken);
+        var botHandler = new BotHandler(botToken, new LogicUpdate(new DbHendler()));
 
         Console.WriteLine("Бот запущен. Нажмите Enter для выхода.");
         Console.ReadLine();
