+++ bot-tr/UserDataList.cs
@@ -1,15 +1,15 @@
-﻿using bot_tr.model;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿//using bot_tr.model;
+//using Newtonsoft.Json;
+//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Text;
+//using System.Threading.Tasks;
 
-namespace bot_tr
-{
-    internal class UserDataList
-    {
-        public List<UserData>? UsersDataList { get; set; }
-    }
-}
+//namespace bot_tr
+//{
+//    internal class UserDataList
+//    {
+//        public List<UserData>? UsersDataList { get; set; }
+//    }
+//}
